Moved the YT downloader initialization into the route to fix a headers being set error cause the object was not destroyed

+++ server.js
@@ -13,15 +13,6 @@ if(!fs.existsSync(`${__dirname}/Videos`)) {
     fs.mkdirSync(`${__dirname}/Videos`);
 }
 
-const YD = new YoutubeToMp3Downloader({
-    "ffmpegPath": "/usr/bin/ffmpeg",        // FFmpeg binary location
-    "outputPath": "./Videos",               // Output file location (default: the home directory)
-    "youtubeVideoQuality": "highestaudio",  // Desired video quality (default: highestaudio)
-    "queueParallelism": 2,                  // Download parallelism (default: 1)
-    "progressTimeout": 2000,                // Interval in ms for the progress reports (default: 1000)
-    "allowWebm": false
-})
-
 app.get('/search/:keywords/:currentPage', async (req, res) => {
     const keywords = req.params.keywords; 
     const currentPage = req.params.currentPage;
@@ -32,20 +23,34 @@ app.get('/search/:keywords/:currentPage', async (req, res) => {
 });
 
 app.get('/video/:id', (req, res) => {
+    const YD = new YoutubeToMp3Downloader({
+        "ffmpegPath": "/usr/bin/ffmpeg",        // FFmpeg binary location
+        "outputPath": "./Videos",               // Output file location (default: the home directory)
+        "youtubeVideoQuality": "highestaudio",  // Desired video quality (default: highestaudio)
+        "queueParallelism": 1,                  // Download parallelism (default: 1)
+        "progressTimeout": 2000,                // Interval in ms for the progress reports (default: 1000)
+        "allowWebm": false
+    })
+
     YD.download(req.params.id);
 
     YD.on("finished", (err, video) => {
-        res.status(200).sendFile(`${__dirname}/Videos/${video.videoTitle}.mp3`, _ => {
-            console.log(`Sent file: ${video.videoTitle}`);
+        res.status(200).json(video);
 
-            try {
-                fs.unlinkSync(`${__dirname}/Videos/${video.videoTitle}.mp3`);
-                console.log(`Deleted file: ${video.videoTitle}`);
-            } catch(err) {
-                console.log(`Couldn't unlink ${video.videoTitle}`);
-                console.log(err);
-            }
-        });
+        if(fs.existsSync(`${__dirname}/Videos/${video.videoTitle}.mp3`)) {
+            fs.unlinkSync(`${__dirname}/Videos/${video.videoTitle}.mp3`);
+            console.log(`Deleted file: ${video.videoTitle}`);
+        }
+
+        // res.status(200).sendFile(`${__dirname}/Videos/${video.videoTitle}.mp3`, _ => {
+        //     console.log(video);
+        //     console.log(`Sent file: ${video.videoTitle}`);
+
+        //     if(fs.existsSync(`${__dirname}/Videos/${video.videoTitle}.mp3`)) {
+        //         fs.unlinkSync(`${__dirname}/Videos/${video.videoTitle}.mp3`);
+        //         console.log(`Deleted file: ${video.videoTitle}`);
+        //     }
+        // });
     });
 
     YD.on("progress", progress => {
@@ -57,7 +62,7 @@ app.get('/video/:id', (req, res) => {
             message: "Couldn't download the file"
         })
     });
-})
+});
 
 app.listen(port, () => {
     console.log(`Example app listening at http://localhost:${port}`)
